requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
# openai==1.12.0 # Commented out or removed
google-generativeai 